Move setAttributes into trait
